2015 - day 7 (elixir)
Memoization seemed to be absolutely essential to have the program
execute in a reasonable amount of time. I started off not having
memoization, and it ran for so long that I couldn't tell if the code was
correct or not. With memoization turned on, it became evident that my
logic was working, it just couldn't get to an answer fast.